rev: update tests for new impl
